chore: move RpgCompanionBottomBar
